Optimizer: convert the RangeDumper pass to a test
